Major spring cleaning!
Gui changes
Removed some warnings
Removed old commented code
Added dragon model without animations
Removed x86 target
Added fbx dll for performance configuration
Removed ununsed assets
Switched out mirror planes for mirror model
Added demo image and report pdf

+++ Project_DXR/src/DX12/Win32Window.cpp
@@ -113,12 +113,12 @@ LRESULT Win32Window::MsgProc(HWND hwnd, UINT msg, WPARAM wParam, LPARAM lParam) 
 		Input::ProcessMessage(msg, wParam, lParam);
 		break;
 	case WM_KEYDOWN:
-		Input::RegisterKeyDown(MapVirtualKeyA(wParam, MAPVK_VK_TO_CHAR));
+		Input::RegisterKeyDown(MapVirtualKeyA((UINT)wParam, MAPVK_VK_TO_CHAR));
 		break;
 	case WM_SYSKEYDOWN:
 		break;
 	case WM_KEYUP:
-		Input::RegisterKeyUp(MapVirtualKeyA(wParam, MAPVK_VK_TO_CHAR));
+		Input::RegisterKeyUp(MapVirtualKeyA((UINT)wParam, MAPVK_VK_TO_CHAR));
 		break;
 	case WM_SYSKEYUP:
 		break;
